+++ src/llm_agents_from_scratch/errors/task_handler.py
@@ -7,3 +7,9 @@ class TaskHandlerError(LLMAgentsFromScratchError):
     """Base error for all TaskHandler-related exceptions."""
 
     pass
+
+
+class RecordMemoryError(TaskHandlerError):
+    """Raised when record_memory() is called with invalid arguments."""
+
+    pass